# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GreenVolt Solar Panel Service Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/AuthUser"
  /auth/push-token:
    post:
      operationId: registerPushToken
      tags: [auth]
      summary: Register or update Expo push token for the current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenInput"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerInput"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffListResponse"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffInput"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all service jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: staffId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of service jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, cancelled, expired]
        - name: customerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListResponse"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, success, failed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create/record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /contact:
    get:
      operationId: listContactSubmissions
      tags: [contact]
      summary: List all contact form submissions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Contact submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contact/{id}/convert:
    post:
      operationId: convertContactToCustomer
      tags: [contact]
      summary: Convert a contact submission into a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertContactInput"
      responses:
        "201":
          description: Customer created from contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Contact not found
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notification history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [service_scheduled, service_completed, subscription_expiry]
        - name: status
          in: query
          schema:
            type: string
            enum: [sent, failed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
...
components:
  schemas:
    AuthUser:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff]
        staffId:
          type: integer
          nullable: true
    PushTokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Expo push token (ExponentPushToken[...])
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        solarCapacity:
          type: number
          nullable: true
        installationDate:
          type: string
          nullable: true
        installationDetails:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - address
        - createdAt
        - updatedAt
    CreateCustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        solarCapacity:
          type: number
        installationDate:
          type: string
        installationDetails:
          type: string
        city:
          type: string
        notes:
          type: string
      required:
        - name
        - phone
        - address
    UpdateCustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        solarCapacity:
          type: number
        installationDate:
          type: string
        installationDetails:
          type: string
        city:
          type: string
        notes:
          type: string
    CustomerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Staff:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        workArea:
          type: string
          nullable: true
        availability:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateStaffInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        workArea:
          type: string
        availability:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - phone
        - role
...
    StaffListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Staff"
        total:
          type: integer
      required:
        - data
        - total
    Service:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        staffId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        scheduledDate:
          type: string
        notes:
          type: string
          nullable: true
        serviceType:
          type: string
          nullable: true
        preServiceImage:
          type: string
          nullable: true
        postServiceImage:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        staff:
          $ref: "#/components/schemas/Staff"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerId
        - status
        - scheduledDate
        - createdAt
        - updatedAt
    CreateServiceInput:
      type: object
      properties:
        customerId:
          type: integer
        staffId:
          type: integer
        scheduledDate:
          type: string
        notes:
          type: string
        serviceType:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
      required:
        - customerId
        - scheduledDate
...
    ServiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Subscription:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        plan:
          type: string
          enum: [basic, standard, premium]
        visitsPerMonth:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, paused, cancelled, expired]
        amount:
          type: number
        daysUntilExpiry:
          type: integer
          nullable: true
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerId
        - plan
        - visitsPerMonth
        - startDate
        - endDate
        - status
        - amount
        - createdAt
        - updatedAt
    CreateSubscriptionInput:
      type: object
      properties:
        customerId:
          type: integer
        plan:
          type: string
          enum: [basic, standard, premium]
        startDate:
          type: string
        amount:
          type: number
      required:
        - customerId
        - plan
        - startDate
        - amount
...
    SubscriptionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Payment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        subscriptionId:
          type: integer
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [pending, success, failed]
        paymentMethod:
          type: string
          nullable: true
        transactionId:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerId
        - amount
        - status
        - createdAt
        - updatedAt
    CreatePaymentInput:
      type: object
      properties:
        customerId:
          type: integer
        subscriptionId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, success, failed]
        paymentMethod:
          type: string
        transactionId:
          type: string
        description:
          type: string
      required:
        - customerId
        - amount
        - status
...
    PaymentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        subject:
          type: string
          nullable: true
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - message
        - isRead
        - createdAt
    ConvertContactInput:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
      required:
        - address
    CreateContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - message
    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [service_scheduled, service_completed, subscription_expiry]
        recipientPhone:
          type: string
        recipientName:
          type: string
          nullable: true
        message:
          type: string
        status:
          type: string
          enum: [sent, failed]
        provider:
          type: string
          nullable: true
        providerMessageId:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        serviceId:
          type: integer
          nullable: true
        subscriptionId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - type
        - recipientPhone
        - message
        - status
        - createdAt
    NotificationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT